refactor:showImageBox  전역 loading 스테이트 삭제, Mutation IsPending 사용으로 변경, generateMutation 부모 컴포넌트로 옮김
